libshared: ensure the backup vars are the same size as what gets copied to it - keep them signed however as it can potentially contain a negative value

+++ release/src/router/shared/misc.c
@@ -1044,8 +1044,8 @@ char *get_productid(void)
 	return productid;
 }
 
-int backup_rx = 0;
-int backup_tx = 0;
+long backup_rx = 0;
+long backup_tx = 0;
 
 unsigned int netdev_calc(char *ifname, char *ifname_desc, unsigned long *rx, unsigned long *tx, char *ifname_desc2, unsigned long *rx2, unsigned long *tx2)
 {
